+++ back/src/main/java/repz/app/persistence/repository/UserRepository.java
@@ -8,12 +8,12 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 
     List<User> findByDeletedAtIsNull();
 
-    Optional<User> findByIdAndDeletedAtIsNull(Integer id);
+    Optional<User> findByIdAndDeletedAtIsNull(Long id);
 
     List<User> findByActiveTrueAndDeletedAtIsNull();
